+++ src/main/java/com/kakao/saramaracommunity/board/service/reqeust/BoardDeleteServiceRequest.java
@@ -1,4 +1,4 @@
-package com.kakao.saramaracommunity.board.dto.business.reqeust;
+package com.kakao.saramaracommunity.board.service.reqeust;
 
 import lombok.Builder;
