+++ src/test/java/com/kakao/saramaracommunity/board/service/BoardServiceImplTest.java
@@ -1,10 +1,10 @@
 package com.kakao.saramaracommunity.board.service;
 
-import com.kakao.saramaracommunity.board.dto.business.reqeust.BoardCreateServiceRequest;
-import com.kakao.saramaracommunity.board.dto.business.reqeust.BoardDeleteServiceRequest;
-import com.kakao.saramaracommunity.board.dto.business.reqeust.BoardUpdateServiceRequest;
-import com.kakao.saramaracommunity.board.dto.business.response.BoardCreateResponse;
-import com.kakao.saramaracommunity.board.dto.business.response.BoardGetResponse;
+import com.kakao.saramaracommunity.board.service.reqeust.BoardCreateServiceRequest;
+import com.kakao.saramaracommunity.board.service.reqeust.BoardDeleteServiceRequest;
+import com.kakao.saramaracommunity.board.service.reqeust.BoardUpdateServiceRequest;
+import com.kakao.saramaracommunity.board.controller.dto.response.BoardCreateResponse;
+import com.kakao.saramaracommunity.board.controller.dto.response.BoardGetResponse;
 import com.kakao.saramaracommunity.board.entity.Board;
 import com.kakao.saramaracommunity.board.entity.BoardImage;
 import com.kakao.saramaracommunity.board.entity.CategoryBoard;
